refactor types import
