+++ frontend/src/pages/edit/EditPage.tsx
@@ -4,11 +4,11 @@ import { BOOK } from '../../api/bookClient';
 import Arrow from '../../assets/icon/arrow-left-solid.svg';
 import { stockData } from '../../model/data';
 import { STATUS } from '../../model/status';
+import { Book, OmitID, TError, ToastContextType } from '../../model/types';
 import { ErrorMessage } from '../../shared/components/error/Error';
 import { BookForm } from '../../shared/components/form/BookForm';
 import { Spinner } from '../../shared/components/spinner/Spinner';
 import { ToastSetState } from '../../shared/context/toastContext';
-import { Book, OmitID, TError, ToastContextType } from '../../types';
 
 export const EditPage = () => {
   const navigate = useNavigate();
