+++ frontend/src/pages/home/HomePage.tsx
@@ -3,9 +3,9 @@ import ReactPaginate from 'react-paginate';
 import { Link } from 'react-router-dom';
 import { BOOK } from '../../api/bookClient';
 import { stockData } from '../../model/data';
+import { Book, PaginatedData, TError } from '../../model/types';
 import { ErrorMessage } from '../../shared/components/error/Error';
 import { Spinner } from '../../shared/components/spinner/Spinner';
-import { Book, PaginatedData, TError } from '../../types';
 import { BookList } from './components/BookList';
 
 export const HomePage: React.FC = () => {
