+++ frontend/src/pages/home/components/BookCard.tsx
@@ -1,5 +1,5 @@
 import { Link } from 'react-router-dom';
-import { Book } from '../../../types';
+import { Book } from '../../../model/types';
 
 type Props = {
   book: Book;
