--- frontend/todo.md
@@ -1,17 +1,0 @@
-# To do:
-
-    1 - Finire lo scheletro - in corso
-    2 - Top bar componente
-
-# Get:
-
-    - Style
-    - Navigazione Aggiungi libro verso la pagina di creazione libro - Fatto
-    - Chiamata Api
-
-# Set:
-
-    - Form creazione libro - Fatto
-    - Funzioni salva e Annulla nel componente top bar
-    - Style
-    - Chiamata APi
